Redux migration in HomePage
