+++ app/reducers/.eslintrc
@@ -1,0 +1,21 @@
+{
+  "env": {
+    "mocha": true
+  },
+  "rules": {
+    "global-require": 0,
+    "consistent-return": 0,
+    "fp/no-class": "error",
+    "fp/no-delete": "error",
+    "fp/no-events": "error",
+    "fp/no-let": "error",
+    "fp/no-loops": "error",
+    "fp/no-mutating-assign": "error",
+    "fp/no-mutation": "error",
+    # "fp/no-nil": "error",
+    "fp/no-proxy": "error",
+    "fp/no-this": "error",
+    "fp/no-throw": "error",
+    "fp/no-unused-expression": "error"
+  }
+}
