+++ test/components/Card.spec_.js
@@ -15,10 +15,13 @@ function setup(propsOverride) {
     genres: ['action', 'comedy'],
     rating: 4.5,
     kind: 'shows',
-    baseUrl: '/item/shows'
+    baseUrl: '/item/shows',
+    type: 'movie'
   }, propsOverride);
 
   const component = renderIntoDocument(<Card {...props} />);
+
+  console.log(component);
 
   return {
     component,
